moved toggleLike into postActions and added conditional REPOST label to post
